fix: skip BNB migration instead of error if no BNB denom present (#1825)
If the BNB denom is not present in the leverage registry, skips the BNB migration instead of causing an error.

This shouldn't affect mainnet or `canon-2`, but is good form in case we ever run this migration on a different testnet.

(cherry picked from commit c5003384209930aa2b6369898c2d13460f43334d)